+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Change the verification of the grid in MAPL_GetGlobalHorzIJIndex to avoid collective call
+- Changed all `MPI_Send` and `MPI_Isend` to `MPI_Ssend` and `MPI_Issend`
 
 ### Fixed
 
@@ -1548,7 +1549,7 @@ new interface with conventional ordering has been introduced.
 ### Changed
 
 - Change to non-blocking send and receive from frontend to beckend in the class MultiGroupServer
-- Change one sided mpi_put to mpi_send and receive pair in the class MultiGroupServer
+- Change one sided mpi_put to MPI_Ssend and receive pair in the class MultiGroupServer
 - Change command line interface to --npes_backend_pernode to avoid confusion
 - Remove self-defined-in-file MAPL macros